+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ Changelog
 
 ## 0.18.2
   * Remove `assign_emodels2`; was a work around for the NCX project
+  * be more relaxed about matching `layer` and `subregion`
+  * for emodel assignment, try `layer` first, and fall back to `subregion` columns
 
 ## 0.18.1
   * Densities are now loaded using float64 precision. Higher precision aids in calculating